chore(cookies): expose splitCookiesString
